requirements.txt
Django==1.8.7
django-extensions==1.5.9
django-ipware==1.1.3
sxclient==0.12.2
PyYAML==3.11
django-sizefield==0.9
user-agents==1.0.1
